Update pkgutil.conf.erb
@juliengk,
The outcome of this change is not what we expect. The use of 'UNSET' is not the best choice in this case, for example with all unset values (default) the pkgutil.conf sets them with blank. I understand why - Take this example;

<% if @catalog_not_cached_real != 'UNSET' -%>
catalog_not_cached=<%= @catalog_not_cached_real %>
<% else -%>
#catalog_not_cached=true
<% end -%>
 

To the template, catalog_not_cached_real is set to "" from the manifests/init.pp
 if $catalog_not_cached != 'UNSET' {
    if type($catalog_not_cached) == 'String' {
      $catalog_not_cached_real = str2bool($catalog_not_cached)  <<<------
    } else {
      $catalog_not_cached_real = $catalog_not_cached
    }
    validate_bool($catalog_not_cached_real)
  }


Instead of 'UNSET' please use undef, which will then be nil - and also update the check accordingly.

For inspiration, please check the current ssh module,